@@ -32,8 +32,14 @@ const targetDir = "./src/content/posts/"
 const fullPath = path.join(targetDir, fileName)
 
 if (fs.existsSync(fullPath)) {
-  console.error(`Error：File ${fullPath} already exists `)
+  console.error(`Error: File ${fullPath} already exists `)
   process.exit(1)
+}
+
+// recursive mode creates multi-level directories
+const dirPath = path.dirname(fullPath)
+if (!fs.existsSync(dirPath)) {
+    fs.mkdirSync(dirPath, { recursive: true })
 }
 
 const content = `---
