@@ -1,8 +1,8 @@
-import type { AstroIntegration } from '@swup/astro'
+import type { AstroIntegration } from "@swup/astro";
 
 declare global {
-  interface Window {
-    // type from '@swup/astro' is incorrect
-    swup: AstroIntegration
-  }
+	interface Window {
+		// type from '@swup/astro' is incorrect
+		swup: AstroIntegration;
+	}
 }
