@@ -1,38 +1,38 @@
-import Key from '../i18nKey'
-import type { Translation } from '../translation'
+import Key from "../i18nKey";
+import type { Translation } from "../translation";
 
 export const ko: Translation = {
-  [Key.home]: '홈',
-  [Key.about]: '소개',
-  [Key.archive]: '아카이브',
-  [Key.search]: '검색',
+	[Key.home]: "홈",
+	[Key.about]: "소개",
+	[Key.archive]: "아카이브",
+	[Key.search]: "검색",
 
-  [Key.tags]: '태그',
-  [Key.categories]: '카테고리',
-  [Key.recentPosts]: '최근 게시물',
+	[Key.tags]: "태그",
+	[Key.categories]: "카테고리",
+	[Key.recentPosts]: "최근 게시물",
 
-  [Key.comments]: '댓글',
+	[Key.comments]: "댓글",
 
-  [Key.untitled]: '제목 없음',
-  [Key.uncategorized]: '분류되지 않음',
-  [Key.noTags]: '태그 없음',
+	[Key.untitled]: "제목 없음",
+	[Key.uncategorized]: "분류되지 않음",
+	[Key.noTags]: "태그 없음",
 
-  [Key.wordCount]: '단어',
-  [Key.wordsCount]: '단어',
-  [Key.minuteCount]: '분',
-  [Key.minutesCount]: '분',
-  [Key.postCount]: '게시물',
-  [Key.postsCount]: '게시물',
+	[Key.wordCount]: "단어",
+	[Key.wordsCount]: "단어",
+	[Key.minuteCount]: "분",
+	[Key.minutesCount]: "분",
+	[Key.postCount]: "게시물",
+	[Key.postsCount]: "게시물",
 
-  [Key.themeColor]: '테마 색상',
+	[Key.themeColor]: "테마 색상",
 
-  [Key.lightMode]: '밝은 모드',
-  [Key.darkMode]: '어두운 모드',
-  [Key.systemMode]: '시스템 모드',
+	[Key.lightMode]: "밝은 모드",
+	[Key.darkMode]: "어두운 모드",
+	[Key.systemMode]: "시스템 모드",
 
-  [Key.more]: '더 보기',
+	[Key.more]: "더 보기",
 
-  [Key.author]: '저자',
-  [Key.publishedAt]: '게시일',
-  [Key.license]: '라이선스',
-}
+	[Key.author]: "저자",
+	[Key.publishedAt]: "게시일",
+	[Key.license]: "라이선스",
+};
