@@ -1,16 +1,16 @@
-import type { APIRoute } from 'astro'
+import type { APIRoute } from "astro";
 
 const robotsTxt = `
 User-agent: *
-Allow: /
+Disallow: /_astro/
 
-Sitemap: ${new URL('sitemap-index.xml', import.meta.env.SITE).href}
-`.trim()
+Sitemap: ${new URL("sitemap-index.xml", import.meta.env.SITE).href}
+`.trim();
 
 export const GET: APIRoute = () => {
-  return new Response(robotsTxt, {
-    headers: {
-      'Content-Type': 'text/plain; charset=utf-8',
-    },
-  })
-}
+	return new Response(robotsTxt, {
+		headers: {
+			"Content-Type": "text/plain; charset=utf-8",
+		},
+	});
+};
