@@ -1,16 +1,17 @@
-import { toString } from 'mdast-util-to-string'
+// biome-ignore lint/suspicious/noShadowRestrictedNames: <explanation>
+import { toString } from "mdast-util-to-string";
 
 /* Use the post's first paragraph as the excerpt */
 export function remarkExcerpt() {
-  return (tree, { data }) => {
-    let excerpt = ''
-    for (let node of tree.children) {
-      if (node.type !== 'paragraph') {
-        continue
-      }
-      excerpt = toString(node)
-      break
-    }
-    data.astro.frontmatter.excerpt = excerpt
-  }
+	return (tree, { data }) => {
+		let excerpt = "";
+		for (const node of tree.children) {
+			if (node.type !== "paragraph") {
+				continue;
+			}
+			excerpt = toString(node);
+			break;
+		}
+		data.astro.frontmatter.excerpt = excerpt;
+	};
 }
